Disable watchdog
This is a bit of a sledge-hammer method, but for now disable
the watchdog, because it kicks in when doing a soft-reboot
or a system kexec call.
Both cut down the reboot time tremendously.
The proper way would be to find out why the watchdog is not
fed in those situations.