Fix ovsdb permissions race
Doing it in the same unit has a race-condition,
as the osdb-server forks of and will create
eventually the socket, and at the same time
the ExecStartPost script is running.
By moving it to its own unit, the script can
wait for the socket and adjust the permissions
without delaying other units.